requirements.txt

Flask==2.1.2
Flask-CORS==3.0.10
pystray==0.16.0
Pillow==9.1.0
pywin32==303
